# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Elliptic Curve Cryptography service API for IOT system",
        "title": "ECC SERVICE API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "NGUYEN LE BAO TRUONG",
            "url": "https://github.com/allvisss",
            "email": "truongnlbse140940@fpt.edu.vn"
        },
        "version": "0.0"
    },
    "paths": {
        "/v1/decrypt": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Decrypt"
                ],
                "summary": "Decrypt ciphertext into plaintext (using curve secp112r1 as demo)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ciphertext",
                        "name": "ciphertext",
                        "in": "query"
                    },
                    {
                        "description": "Keypair",
                        "name": "priv",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Keypair"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/decrypt2": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Decrypt"
                ],
                "summary": "Decrypt ciphertext into plaintext",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Ciphertext",
                        "name": "ciphertext",
                        "in": "query"
                    },
                    {
                        "description": "Keypair",
                        "name": "priv",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Keypair"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/encrypt": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Encrypt"
                ],
                "summary": "Encrypt plaintext into ciphertext (using curve secp112r1 as demo)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Plaintext",
                        "name": "msg",
                        "in": "query"
                    },
                    {
                        "description": "Keypair",
                        "name": "priv",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Keypair"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/encrypt2": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Encrypt"
                ],
                "summary": "Encrypt plaintext into ciphertext",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Plaintext",
                        "name": "msg",
                        "in": "query"
                    },
                    {
                        "description": "Key",
                        "name": "priv",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/api.Keypair"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/keygen": {
            "get": {
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Key"
                ],
                "summary": "Generate key pair (using curve secp112r1 as demo)",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/keygen2": {
            "get": {
                "consumes": [
                    "*/*"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Key"
                ],
                "summary": "Generate key pair",
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/v1/testDecrypt": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Test"
                ],
                "summary": "Testing decrypt method",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Plaintext",
                        "name": "plaintext",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "api.Keypair": {
            "type": "object",
            "properties": {
                "D": {
                    "type": "string"
                },
                "X": {
                    "type": "string"
                },
                "Y": {
                    "type": "string"
                },
                "curve": {
                    "type": "object",
                    "properties": {
                        "B": {
                            "type": "string"
                        },
                        "BitSize": {
                            "type": "integer"
                        },
                        "Gx": {
                            "type": "string"
                        },
                        "Gy": {
                            "type": "string"
                        },
                        "N": {
                            "type": "string"
                        },
                        "Name": {
                            "type": "string"
                        },
                        "P": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
        "response.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {
                    "type": "object"
                }
            }
        }
    }
}